pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-dcip"
version = "0.1.0"
description = "Django app for detecting datacenter/VPN IP addresses"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "django-cidrfield",
    "netaddr",
]

[project.urls]
Homepage = "https://github.com/NameOcean/dcip"

[tool.setuptools.packages.find]
include = ["dcip*"]
